renamed the portrait and landscape composables for single counter.
