+++ app/src/main/java/com/example/stitchcounterv3/feature/single/SingleCounterLandscapeLayout.kt
@@ -10,6 +10,7 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Surface
@@ -21,17 +22,16 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.stitchcounterv3.feature.sharedComposables.AdjustmentButtons
+import com.example.stitchcounterv3.domain.model.AdjustmentAmount
 import com.example.stitchcounterv3.ui.theme.StitchCounterV3Theme
 
 @Composable
-fun SingleCounterPortraitScreen(
+fun SingleCounterLandScapeLayout(
     state: SingleCounterUiState,
     viewModel: SingleCounterViewModel
 ) {
-    Column(
-        modifier = Modifier.padding(24.dp),
-        verticalArrangement = Arrangement.spacedBy(16.dp)
+    Row(
+        modifier = Modifier.padding(24.dp)
     ) {
         Text(
             "Basic Counter",
@@ -72,10 +72,28 @@ fun SingleCounterPortraitScreen(
                 )
             }
         }
-        AdjustmentButtons(
-            onAdjustmentClick = { viewModel.changeAdjustment(it) },
-            modifier = Modifier.fillMaxWidth()
-        )
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.spacedBy(12.dp),
+        ) {
+            Spacer(modifier = Modifier.weight(1f))
+            Button(
+                colors = ButtonDefaults.buttonColors().copy(containerColor = MaterialTheme.colorScheme.secondary),
+                onClick = { viewModel.changeAdjustment(AdjustmentAmount.ONE) }) {
+                Text("+1")
+            }
+            Button(
+                colors = ButtonDefaults.buttonColors().copy(containerColor = MaterialTheme.colorScheme.tertiary),
+                onClick = { viewModel.changeAdjustment(AdjustmentAmount.FIVE) }) {
+                Text("+5")
+            }
+            Button(
+                colors = ButtonDefaults.buttonColors().copy(containerColor = MaterialTheme.colorScheme.error),
+                onClick = { viewModel.changeAdjustment(AdjustmentAmount.TEN) }) {
+                Text("+10")
+            }
+            Spacer(modifier = Modifier.weight(1f))
+        }
         Row(
             modifier = Modifier.fillMaxWidth(),
             horizontalArrangement = Arrangement.Center
@@ -96,7 +114,7 @@ fun SingleCounterPortraitScreen(
 
 @Preview
 @Composable
-private fun SingleCounterPortraitPreview() {
+private fun SingleCounterLandScapeScreenPreview() {
     StitchCounterV3Theme {
         // Preview without navigation dependencies
         Surface(modifier = Modifier.fillMaxSize()) {
